refactor: update ProfileModal import paths and simplify modal usage across multiple pages
