+++ frontend_web/react+vite/src/modules/warranties/WarrantiesPage.jsx
@@ -9,12 +9,12 @@ import TopSection from "../../globals/components/ui/TopSection";
 import WarrantiesTable from "./components/ui/WarrantiesTable";
 // Modales
 import Modal from "../../globals/components/modals/Modal";
-import FilterModal from "../../globals/components/modals/FilterModal";
-import ProfileModal from "../../globals/components/modals/ProfileModal";
+import MoreWarrantyInfo from "./components/modals/MoreWarrantyInfo";
 import AddWarrantyModal from "./components/modals/AddWarrantyModal";
 import EditWarrantyModal from "./components/modals/EditWarrantyModal";
+import FilterModal from "../../globals/components/modals/FilterModal";
 import DeleteWarrantyModal from "./components/modals/DeleteWarrantyModal";
-import MoreWarrantyInfo from "./components/modals/MoreWarrantyInfo";
+import ProfileModal from "../../globals/components/modals/profileModal/ProfileModal";
 
 export default function WarrantiesPage() {
   const { isOpen, modalData, modalType, refetch, openModal, closeModal } =
@@ -56,7 +56,7 @@ export default function WarrantiesPage() {
           isOpen={isOpen}
           onClose={closeModal}
         >
-          {modalType === "user" && <ProfileModal onClose={closeModal} />}
+          {modalType === "user" && <ProfileModal />}
           {modalType === "filter" && <FilterModal onClose={closeModal} />}
           {modalType === "add" && (
             <AddWarrantyModal
